+++ modules/core/dependency/python-ihm/make-release.sh
@@ -1,6 +1,7 @@
 #!/bin/bash -e
 
 # First, do
+# - Update AuditConformDumper to match latest IHM dictionary if necessary
 # - Update ChangeLog.rst with the release number
 # - Update release number in ihm/__init__.py, MANIFEST.in, and setup.py
 # - Commit, tag, and push